requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.1.0.post1
fastapi==0.110.0
filelock==3.13.1
gdown==5.1.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
orjson==3.9.15
packaging==23.2
pandas==2.2.1
psycopg2-binary==2.9.9
pydantic==2.6.2
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.2.2
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
threadpoolctl==3.3.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
unrar==0.4
unzip==1.0.0
urllib3==2.2.1
uvicorn==0.27.1
waitress==3.0.0
watchfiles==0.21.0
websockets==12.0
